Updated earning status from db

+++ client/app/src/main/java/com/microsoft/research/karya/data/model/karya/modelsExtra/EarningStatus.kt
@@ -1,7 +1,7 @@
 package com.microsoft.research.karya.data.model.karya.modelsExtra
 
 data class EarningStatus(
-  val earnedWeek: Int,
-  val earnedTotal: Int,
-  val paidTotal: Int
+  val earnedWeek: Float,
+  val earnedTotal: Float,
+  val paidTotal: Float
 )
